refactor(arch): decouple metal backend from kernel — part 1 (shared state + layout)
Cut the easy upward references from the bare-metal arch backend into the kernel,
in preparation for lifting `kernel/src/arch/` into its own `arch-metal` crate
(sibling to `arch-interp`):

- `PM_STEP_BUDGET` (single-step trace budget) moves to `arch-abi`: the backend's
  `#DB` handler consumes it, the DOS/DPMI layer only arms it, and both depend on
  `arch-abi`. DOS re-exports it so `dos::PM_STEP_BUDGET` sites are unchanged.
- The kernel-heap VA window (`HEAP_END` + `heap_base()`) moves to `arch::paging2`
  as a memory-layout fact (the demand-paging `#PF` handler grows it). The
  metal-only heap allocator now reads it from arch; the `traps` `#PF` handler
  uses the local copy instead of reaching into `crate::kernel::heap`.
- Arch type references (`UserMode`/`Regs`/`Frame64`) repointed from the kernel's
  re-exports to `arch_abi` directly; `RootPageTable` to the arch-local path.

Remaining arch→kernel coupling (next parts): the `pm_step_log` tracer call, the
`println!`/`dbg_println!` logging macros, and `boot.rs` (the metal entry, which
becomes the binary). Both backends build.
