+++ kernel/src/arch/mod.rs
@@ -23,7 +23,7 @@ mod x86;
 
 // Types
 pub use backend::Metal;
-pub use paging2::{KernelPages, RawPage, RootPageTable, PAGE_SIZE, LOW_MEM_BASE};
+pub use paging2::{KernelPages, RawPage, RootPageTable, PAGE_SIZE, LOW_MEM_BASE, HEAP_END, heap_base};
 pub use vcpu::{Vcpu, set_current_vcpu};
 pub use irq::Irq;
 pub use descriptors::{USER_CS, USER_CS64, USER_DS};
